bots/picker_bot/requirements.txt
aiogram>=3.13
python-dotenv
aiohttp
